Altering Main functionality
Moved the napari GUI stuff to the main script because it was not working correctly as a class object. I also am working on a new functionality in how to draw the ellipse. It was not able to be rotated before, so now the ellipse is generated by fitting an ellipse to a user-created polygon. This is not done yet though